--- alchemiscale_fah/tests/integration/compute/test_client.py
@@ -1,20 +1,0 @@
-"""Tests for FahAdaptiveSamplingClient.
-
-"""
-
-import pytest
-
-
-class TestFahAdaptiveSamplingClient:
-
-    def test_create_clone(client):
-        ...
-
-    def test_get_clone():
-        ...
-
-    def test_create_clone_file():
-        ...
-
-    def test_create_clone_file_from_bytes():
-        ...
