--- alchemiscale_fah/tests/integration/compute/test_service.py
@@ -1,3 +1,0 @@
-"""Tests for compute services.
-
-"""
